This is the start of code that will allow us to register new users and also force users who don't have a default customer profile to add a new one.
With this I'm going to knock off day two. Day two was a bit
disappointing, I didn't get many new features implemented - although I
think the rsync stuff is pretty much finished; we can now buy packages
from the front-end; and I've started on creating and viewing (and
sending) invoices. Which I supposed isn't bad, but there was a lot of
refactoring and architectural changes in there.

Key things tomorrow will be to expose subscription cancellation to the
web site (not much work), finishing off the code for registering new
users and customers, (the present commit) allowing users to add a new
customer profile, and working on invoice payment processing. That's
all still pretty infrastructury, but I think after that it's features
all the way, and adding support for databases will be the first one to
go.